+++ qmi/core/transport.py
@@ -803,7 +803,6 @@ class QMI_UsbTmcTransport(QMI_Transport):
         self.vendorid = vendorid
         self.productid = productid
         self.serialnr = serialnr
-        self._read_buffer = bytes()
 
     @staticmethod
     def _validate_product_id(productid):
@@ -816,25 +815,20 @@ class QMI_UsbTmcTransport(QMI_Transport):
             raise QMI_TransportDescriptorException("Missing/bad USB vendor ID {}".format(vendorid))
 
     def __str__(self) -> str:
-        return "QMI_UsbTmcTransport 0x{:04x}:0x{:04x} ({})".format(self.vendorid, self.productid, self.serialnr)
-
-    def read_until_timeout(self, nbytes: int, timeout: float) -> bytes:
-        raise NotImplementedError("QMI_UsbTmcTransport.discard_read not implemented")
-
-    def discard_read(self) -> None:
-        pass
+        return f"QMI_UsbTmcTransport 0x{self.vendorid:04x}:0x{self.productid:04x} ({self.serialnr})"
 
     def close(self) -> None:
         _logger.debug("Closing USBTMC device 0x%04x:0x%04x (%s)", self.vendorid, self.productid, self.serialnr)
         super().close()
 
     def write(self, data: bytes) -> None:
-        """Send the specified data to the instrument as a single USBTMC message.
+        """Send the specified data to the instrument as a single USBTMC message. A fixed timeout of 5 seconds
+        is applied. If timeout occurs, the transfer is aborted and an exception is raised.
 
-        A fixed timeout of 5 seconds is applied.
-        If timeout occurs, the transfer is aborted and an exception is raised.
+        Parameters:
+            data: The data to write in a byte string.
         """
-        raise NotImplementedError("QMI_Transport.read not implemented")
+        raise NotImplementedError("QMI_UsbTmcTransport.write not implemented")
 
     def read(self, nbytes: int, timeout: Optional[float]) -> bytes:
         """Read a specified number of bytes from the transport.
@@ -846,7 +840,7 @@ class QMI_UsbTmcTransport(QMI_Transport):
         data is discarded and QMI_TimeoutException is raised.
 
         Parameters:
-            nbytes: Number of bytes to read.
+            nbytes: Expected number of bytes to read.
             timeout: Maximum time to wait in seconds (default: 60 seconds).
 
         Returns:
@@ -858,50 +852,55 @@ class QMI_UsbTmcTransport(QMI_Transport):
             ~qmi.core.exceptions.QMI_EndOfInputException: If an end-of-message indicator
                 is received before the requested number of bytes is reached.
         """
-
         self._check_is_open()
-
         # USB requires a timeout
         if timeout is None:
             timeout = self.DEFAULT_READ_TIMEOUT
 
-        if not self._read_buffer:
-            # Read buffer is empty - read a new message from the instrument.
-            self._read_buffer = self._read_message(timeout)
+        # Read a USB message
+        ret = self._read_message(timeout)
+        if len(ret) == nbytes:
+            return ret
 
-        # Check that the buffer contains enough data.
-        if len(self._read_buffer) < nbytes:
-            raise QMI_EndOfInputException("Expecting {} bytes but only {} bytes remaining in USBTMC message"
-                                          .format(nbytes, len(self._read_buffer)))
-
-        data = self._read_buffer[:nbytes]
-        self._read_buffer = self._read_buffer[nbytes:]
-        return data
+        _logger.debug("USBTMC read message contained data %s" % ret.decode())
+        raise QMI_EndOfInputException(
+            f"The read message did not contain expected bytes of data ({len(ret)} != {nbytes}."
+        )
 
     def read_until(self, message_terminator: bytes, timeout: Optional[float]) -> bytes:
-        """Read a single USBTMC message from the instrument."
+        """The specified message_terminator is ignored. Instead, the instrument must autonomously indicate the end
+        of the message according to the USBTMC protocol.
 
-        The specified message_terminator is ignored.
-        Instead, the instrument must autonomously indicate the end of the
-        message according to the USBTMC protocol.
+        As the message_terminator is not used, we forward simply the call to the `read_until_timeout` call.
 
-        This method blocks until a complete USBTMC message is available.
+        Parameters:
+            message_terminator: This input is ignored.
+            timeout: Maximum time to wait (in seconds).
+
+        Returns:
+            Received bytes.
+        """
+        return self.read_until_timeout(0, timeout)
+
+    def read_until_timeout(self, nbytes: int, timeout: float) -> bytes:
+        """Read a single USBTMC message from the instrument.
+
+        If there is already data in the buffer, return it.
 
         If the timeout expires before the message is received, the read is
         aborted and any data already received are discarded. In this
         case QMI_TimeoutException is raised.
 
         Parameters:
-            message_terminator: Message terminator characters.
+            nbytes: This input is ignored.
             timeout: Maximum time to wait (in seconds).
 
         Returns:
-            Received bytes, including the terminator.
+            Received bytes.
 
         Raises:
             ~qmi.core.exceptions.QMI_TimeoutException: If the timeout expires before the
-                requested number of bytes are available.
-
+            requested number of bytes are available.
         """
         self._check_is_open()
 
@@ -909,19 +908,28 @@ class QMI_UsbTmcTransport(QMI_Transport):
         if timeout is None:
             timeout = self.DEFAULT_READ_TIMEOUT
 
-        if self._read_buffer:
-            # Data already available in buffer; return it now.
-            data = self._read_buffer
-            self._read_buffer = bytes()
-        else:
-            # Read buffer is empty - read a new message from the instrument.
-            data = self._read_message(timeout)
+        # Read a new message from the instrument.
+        data = self._read_message(timeout)
 
         return data
 
+    def discard_read(self) -> None:
+        try:
+            self._read_message(0.0)
+        except QMI_TimeoutException:
+            return  # Nothing was in the instrument buffer, so we just continue.
+
     def _read_message(self, timeout):
-        """Read one USBTMC message from the instrument."""
-        raise NotImplementedError("QMI_Transport.read not implemented")
+        """Read one USBTMC message from the instrument.
+        This should be implemented in deriving classes.
+
+        Parameters:
+            timeout: Obligatory parameter to set the timeout for reading.
+
+        Raises:
+            QMI_TimeoutException: If the read is not done within timeout period.
+        """
+        raise NotImplementedError("QMI_UsbTmcTransport.read not implemented")
 
     @staticmethod
     def _format_resources(resources: List[str]) -> List[str]:
@@ -942,7 +950,13 @@ class QMI_UsbTmcTransport(QMI_Transport):
 
     @staticmethod
     def list_resources() -> List[str]:
-        raise NotImplementedError("QMI_Transport.read not implemented")
+        """List available resources from USB connections.
+        This should be implemented in deriving classes.
+
+        Returns:
+            resources: List of available resources in QMI's USBTMC transport string format.
+        """
+        raise NotImplementedError("QMI_UsbTmcTransport.list_resources not implemented")
 
 
 class QMI_Vxi11Transport(QMI_Transport):